user activities

+++ App/app/src/main/java/com/example/chilltime/AdapterMovies.java
@@ -35,6 +35,7 @@ import com.squareup.picasso.Picasso;
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
@@ -45,16 +46,16 @@ import java.util.Map;
 public class AdapterMovies extends RecyclerView.Adapter<AdapterMovies.ViewHolder> {
     List<String> names;
     List<String> images;
-    List<Integer> ids;
+    List<Long> ids;
     Context context;
     LayoutInflater inflater;
     // User
     FirebaseAuth mAuth;
     FirebaseFirestore mStore;
     String userID;
-    List<Integer> favoritesUser;
+    List<Long> favoritesUser;
     List<String> imagesFavoritesUser;
-    List<Integer> watchesUser;
+    List<Long> watchesUser;
     List<String> imagesWatchesUser;
     DocumentReference documentReference;
     long timeWatches;
@@ -64,7 +65,7 @@ public class AdapterMovies extends RecyclerView.Adapter<AdapterMovies.ViewHolder
     // Para mandar o id para outra activity(details)
     public static final String EXTRA_MESSAGE = "com.example.chilltime.extra.MESSAGE";
 
-    public AdapterMovies(Context context, List<String> names, List<String> images, List<Integer> ids, List<Integer> idsFavorites, List<String> imagesFavoritesUser, List<Integer> idsWatches, List<String> imagesWatchesUser, long timeWatches){
+    public AdapterMovies(Context context, List<String> names, List<String> images, List<Long> ids, List<Long> idsFavorites, List<String> imagesFavoritesUser, List<Long> idsWatches, List<String> imagesWatchesUser, long timeWatches){
         this.names = names;
         this.context = context;
         this.images = images;
@@ -96,10 +97,11 @@ public class AdapterMovies extends RecyclerView.Adapter<AdapterMovies.ViewHolder
     @Override
     public void onBindViewHolder(@NonNull final ViewHolder holder, final int position) {
         // verifica se o filme ja esta na bd do user e muda a imagem por ja adicionado e visto
-        if(favoritesUser.contains(ids.get(position).longValue())){
+
+        if(favoritesUser.contains(ids.get(position))){
             holder.addFavorite.setImageResource(R.drawable.removefavorite);
         }
-        if(watchesUser.contains(ids.get(position).longValue())){
+        if(watchesUser.contains(ids.get(position))){
             holder.addWatch.setImageResource(R.drawable.removewatch);
         }
